Implementación Mercado Pago funcional
Mercado Pago funcionando con usuarios de prueba

+++ ms_cart/internal/models/order.go
@@ -9,7 +9,7 @@ type Order struct {
     gorm.Model 
 	
 	UserID   uint      `gorm:"not null;index"` 
-	Total       float64   `gorm:"type:numeric"`
+	Total       int64   `gorm:"type:numeric"`
 	Status      string    `gorm:"default:'PENDING'"`
 	ShippingAddress string   `gorm:"type:text;not null"`
 	OrderItems  []OrderItem `gorm:"foreignKey:OrderID"`
